+++ rtsp-server/net/BufferWriter.h
@@ -26,8 +26,8 @@ public:
 	BufferWriter(int capacity = kMaxQueueLength);
 	~BufferWriter() {}
 
-	bool Append(std::shared_ptr<char> data, uint32_t size, uint32_t index=0);
-	bool Append(const char* data, uint32_t size, uint32_t index=0);
+	bool Append(std::shared_ptr<char> data, size_t size, uint32_t index=0);
+	bool Append(const char* data, size_t size, uint32_t index=0);
 	int Send(SOCKET sockfd, int timeout=0);
 
 	bool IsEmpty() const 
@@ -43,7 +43,7 @@ private:
 	typedef struct 
 	{
 		std::shared_ptr<char> data;
-		uint32_t size;
+		size_t size;
 		uint32_t writeIndex;
 	} Packet;
