+++ rtsp-server/xop/MediaSession.h
@@ -76,10 +76,10 @@ public:
 
     uint16_t GetMulticastPort(MediaChannelId channel_id) const
     {
-		if (channel_id >= multicast_port_.size()) {
+	    if (static_cast<int>(channel_id) >= multicast_port_.size()) {
 			return 0;
 		}         
-        return multicast_port_[channel_id];
+        return multicast_port_[static_cast<int>(channel_id)];
     }
 
 private:
